+++ mjml/tools.py
@@ -1,3 +1,4 @@
+import copy
 import socket
 import random
 import subprocess
@@ -5,15 +6,24 @@ from django.utils.encoding import force_str
 from . import settings as mjml_settings
 
 
+_cache = {}
+
+
 def _mjml_render_by_cmd(mjml_code):
-    cmd_args = mjml_settings.MJML_EXEC_CMD
-    if not isinstance(cmd_args, list):
-        cmd_args = [cmd_args]
-        cmd_args.extend(['-i', '-s'])
+    if 'cmd_args' not in _cache:
+        cmd_args = copy.copy(mjml_settings.MJML_EXEC_CMD)
+        if not isinstance(cmd_args, list):
+            cmd_args = [cmd_args]
+        for ca in ('-i', '-s'):
+            if ca not in cmd_args:
+                cmd_args.append(ca)
+        _cache['cmd_args'] = cmd_args
+    else:
+        cmd_args = _cache['cmd_args']
 
     try:
-        p = subprocess.Popen(cmd_args, stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-        html = force_str(p.communicate(mjml_code.encode('utf8'))[0])
+        p = subprocess.Popen(cmd_args, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        stdout, stderr = p.communicate(mjml_code.encode('utf8'))
     except (IOError, OSError) as e:
         raise RuntimeError(
             'Problem to run command "{}"\n'.format(' '.join(cmd_args)) +
@@ -21,7 +31,10 @@ def _mjml_render_by_cmd(mjml_code):
             'Check that mjml is installed and allow permissions for execute.\n' +
             'See https://github.com/mjmlio/mjml#installation'
         )
-    return html
+    if stderr:
+        raise RuntimeError('MJML stderr is not empty: {}.'.format(force_str(stderr)))
+
+    return force_str(stdout)
 
 
 def _mjml_render_by_tcpserver(mjml_code):