+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.5.0 (2018-04-24)
+==================
+  * Add support MJML v4
+  * Tcpserver doesn't skip mjml errors now
+  * Refactor arguments in tcpserver
+
 0.4.0 (2018-01-10)
 ==================
   * Add support Django 2.0